rpmsg-proto: Avoid deletion of bound socket channels from radix tree.
A radix tree is used by the rpmsg socket driver to keep track of created
channels.

The only channels inserted into the tree are from a probe, which are only
those which can be connected.

The socket bind operation does not result in a probe, but the release of a
bound socket was causing a deletion from the radix tree, causing an error
message.

Bound sockets are only for src addresses below the RESERVED number (1024).

Now, bind operations will avoid deleting from the radix tree, by checking
the src address value first.

Signed-off-by: G Anthony <a0783926@ti.com>