davinci_remoteproc: Loop to handle all incoming rpmsgs in the work function.
This remoteproc implementation uses a workqueue to defer vring_interrupt
handling to a shared global kernel worker thread.

Previously, rapid successive interrupts from the remote processor were
resulting in messages not getting handled, as the schedule_work() function
would sometimes fail to schedule if work was already pending.

This fix essentially loops to handle all pending incoming messages in the
work function.

Since the schedule_work() function clears the workqueue's pending bit flag
prior to calling the work function, any interrupts received during the work
function result in a new work function to be scheduled, which may or may not
be redundant.

Also, added a flush of the workqueue on rproc_stop().

This model is similar to what is done in arch/arm/plat-omap/mailbox.c for the
OMAP mailbox rx interrupt handling.

Signed-off-by: G Anthony <a0783926@ti.com>